Update MAX32650 SDHC_FAT example to use new CLI library (pt. 1)
