+++ Examples/MAX32650/SDHC_FAT/project.mk
@@ -22,3 +22,6 @@ SBT=0
 LIB_SDHC=1
 # Use FatFS R0.15
 FATFS_VERSION = ff15
+
+# Enable CLI Library
+LIB_CLI=1
